refactor: decouple reference harness from main cli
(cherry picked from commit 32fd5cdb92e7393d92fea117d7c33c48dafee4da)
